Initial commit:
right now we're encrypting (and marking) a semi-covert payload into an ICMP
echo-request (ping) and pulling it out on the other side in a shell. the shell
is non-persistant at the moment (really fun issues with piping and buffering
and whatnot), and the ping payload is less than optimal, but whatever, it's
better than nothing.

Note the time being 3 in the freaking morning.